test(cii): move internal docs drift guard
